+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#4418 Fix search when there are multiple vector fields
 -   \#4488 get_entity_by_id() performance is poor in 0.11.0
 -   \#4511 Insert should be failed if field type not matched
+-   \#4536 restful api /collections/{collection_name}/partitions (DELETE) didn't work
 
 ## Feature
 -   \#4163 Update C++ sdk search interface